[types/vm] Move `vm_errors` to `move-core-types` crate.
This allows us to remove the dependency between the `vm` crate and
`libra-types`.
