+++ language/vm/src/file_format.rs
@@ -30,14 +30,12 @@ use crate::{
     access::ModuleAccess, check_bounds::BoundsChecker, errors::VMResult, internals::ModuleIndex,
     IndexKind, SignatureTokenKind,
 };
-use libra_types::{
-    account_address::AccountAddress,
-    vm_error::{StatusCode, VMStatus},
-};
 use mirai_annotations::*;
 use move_core_types::{
+    account_address::AccountAddress,
     identifier::{IdentStr, Identifier},
     language_storage::ModuleId,
+    vm_error::{StatusCode, VMStatus},
 };
 use num_variants::NumVariants;
 #[cfg(any(test, feature = "fuzzing"))]
