+++ language/vm/src/proptest_types/constants.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::file_format::{Constant, SignatureToken};
-use libra_types::account_address::AccountAddress;
+use move_core_types::account_address::AccountAddress;
 use proptest::{
     arbitrary::any,
     collection::{btree_set, vec, SizeRange},
