+++ language/vm/src/unit_tests/deserializer_tests.rs
@@ -5,7 +5,7 @@ use crate::{
     file_format::{CompiledModule, CompiledScript},
     file_format_common::*,
 };
-use libra_types::vm_error::StatusCode;
+use move_core_types::vm_error::StatusCode;
 
 #[test]
 fn malformed_simple() {
