+++ types/src/proptest_types.rs
@@ -478,35 +478,6 @@ impl TransactionPayload {
     }
 }
 
-/// The `Arbitrary` impl only generates validation statuses since the full enum is too large.
-impl Arbitrary for StatusCode {
-    type Parameters = ();
-    type Strategy = BoxedStrategy<Self>;
-
-    fn arbitrary_with(_args: ()) -> Self::Strategy {
-        prop_oneof![
-            Just(StatusCode::UNKNOWN_VALIDATION_STATUS),
-            Just(StatusCode::INVALID_SIGNATURE),
-            Just(StatusCode::INVALID_AUTH_KEY),
-            Just(StatusCode::SEQUENCE_NUMBER_TOO_OLD),
-            Just(StatusCode::SEQUENCE_NUMBER_TOO_NEW),
-            Just(StatusCode::INSUFFICIENT_BALANCE_FOR_TRANSACTION_FEE),
-            Just(StatusCode::TRANSACTION_EXPIRED),
-            Just(StatusCode::SENDING_ACCOUNT_DOES_NOT_EXIST),
-            Just(StatusCode::REJECTED_WRITE_SET),
-            Just(StatusCode::INVALID_WRITE_SET),
-            Just(StatusCode::EXCEEDED_MAX_TRANSACTION_SIZE),
-            Just(StatusCode::UNKNOWN_SCRIPT),
-            Just(StatusCode::UNKNOWN_MODULE),
-            Just(StatusCode::MAX_GAS_UNITS_EXCEEDS_MAX_GAS_UNITS_BOUND),
-            Just(StatusCode::MAX_GAS_UNITS_BELOW_MIN_TRANSACTION_GAS_UNITS),
-            Just(StatusCode::GAS_UNIT_PRICE_BELOW_MIN_BOUND),
-            Just(StatusCode::GAS_UNIT_PRICE_ABOVE_MAX_BOUND),
-        ]
-        .boxed()
-    }
-}
-
 prop_compose! {
     fn arb_transaction_status()(vm_status in any::<VMStatus>()) -> TransactionStatus {
         vm_status.into()
